Update routing to split awards into its own spot
Trying to Prevent website section from being messy
