+++ sites/diverseeducation.com/server/routes/awards.js
@@ -1,0 +1,7 @@
+const emergingScholars = require('./awards/emerging-scholars');
+
+
+module.exports = (app) => {
+  // Emerging Scholars
+  emergingScholars(app);
+};
